+++ lecture/[리액트 2부] 고급 주제와 훅(by 김정환)/1-react/src/pages/ProductPage/index.jsx
@@ -1,0 +1,25 @@
+import Navbar from "../../components/Navbar";
+import Page from "../../components/Page";
+import ProductItem from "../../components/ProductItem";
+import Title from "../../components/Title";
+
+const fakeProduct = {
+  id: "CACDA421",
+  name: "해물 계란 라면",
+  price: "6000",
+  thumbnail: "./images/menu-해물계란라면.jpg",
+};
+
+const ProductPage = () => (
+  <div className="ProductPage">
+    <Page header={<Title>메뉴목록</Title>} footer={<Navbar />}>
+      <ul>
+        <li>
+          <ProductItem product={fakeProduct} />
+        </li>
+      </ul>
+    </Page>
+  </div>
+);
+
+export default ProductPage;
